+++ x/authority/types/message_update_chain_info_test.go
@@ -19,29 +19,17 @@ func TestMsgUpdateChainInfo_ValidateBasic(t *testing.T) {
 	}{
 		{
 			name: "valid message",
-			msg:  types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.ChainInfo(42)),
+			msg:  types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.Chain(42)),
 		},
 		{
 			name:        "invalid creator address",
-			msg:         types.NewMsgUpdateChainInfo("invalid", sample.ChainInfo(42)),
+			msg:         types.NewMsgUpdateChainInfo("invalid", sample.Chain(42)),
 			errContains: "invalid creator address",
 		},
 		{
-			name: "invalid chain info",
-			msg: types.NewMsgUpdateChainInfo(sample.AccAddress(), types.ChainInfo{
-				Chains: []chains.Chain{
-					{
-						ChainId:     0,
-						Network:     chains.Network_optimism,
-						NetworkType: chains.NetworkType_testnet,
-						Vm:          chains.Vm_evm,
-						Consensus:   chains.Consensus_op_stack,
-						IsExternal:  true,
-						Name:        "foo",
-					},
-				},
-			}),
-			errContains: "invalid chain info",
+			name:        "invalid chain",
+			msg:         types.NewMsgUpdateChainInfo(sample.AccAddress(), chains.Chain{ChainId: -1}),
+			errContains: "invalid chain",
 		},
 	}
 
@@ -66,12 +54,12 @@ func TestMsgUpdateChainInfo_GetSigners(t *testing.T) {
 	}{
 		{
 			name:   "valid signer",
-			msg:    types.NewMsgUpdateChainInfo(signer, sample.ChainInfo(42)),
+			msg:    types.NewMsgUpdateChainInfo(signer, sample.Chain(42)),
 			panics: false,
 		},
 		{
 			name:   "invalid signer",
-			msg:    types.NewMsgUpdateChainInfo("invalid", sample.ChainInfo(42)),
+			msg:    types.NewMsgUpdateChainInfo("invalid", sample.Chain(42)),
 			panics: true,
 		},
 	}
@@ -91,17 +79,17 @@ func TestMsgUpdateChainInfo_GetSigners(t *testing.T) {
 }
 
 func TestMsgUpdateChainInfo_Type(t *testing.T) {
-	msg := types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.ChainInfo(42))
+	msg := types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.Chain(42))
 	require.Equal(t, types.TypeMsgUpdateChainInfo, msg.Type())
 }
 
 func TestMsgUpdateChainInfo_Route(t *testing.T) {
-	msg := types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.ChainInfo(42))
+	msg := types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.Chain(42))
 	require.Equal(t, types.RouterKey, msg.Route())
 }
 
 func TestMsgUpdateChainInfo_GetSignBytes(t *testing.T) {
-	msg := types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.ChainInfo(42))
+	msg := types.NewMsgUpdateChainInfo(sample.AccAddress(), sample.Chain(42))
 	require.NotPanics(t, func() {
 		msg.GetSignBytes()
 	})