+++ zetaclient/chains/bitcoin/observer/witness.go
@@ -67,13 +67,20 @@ func GetBtcEventWithWitness(
 		return nil, errors.Wrapf(err, "error getting sender address for inbound: %s", tx.Txid)
 	}
 
+	// skip this tx and move on (e.g., due to unknown script type)
+	// we don't know whom to refund if this tx gets reverted in zetacore
+	if fromAddress == "" {
+		return nil, nil
+	}
+
 	return &BTCInboundEvent{
-		FromAddress: fromAddress,
-		ToAddress:   tssAddress,
-		Value:       amount,
-		MemoBytes:   memo,
-		BlockNumber: blockNumber,
-		TxHash:      tx.Txid,
+		FromAddress:  fromAddress,
+		ToAddress:    tssAddress,
+		Value:        amount,
+		DepositorFee: depositorFee,
+		MemoBytes:    memo,
+		BlockNumber:  blockNumber,
+		TxHash:       tx.Txid,
 	}, nil
 }