+++ zetaclient/chains/bitcoin/signer/signer.go
@@ -9,12 +9,13 @@ import (
 	"math/big"
 	"time"
 
-	"github.com/btcsuite/btcd/btcec"
+	"github.com/btcsuite/btcd/btcec/v2"
+	btcecdsa "github.com/btcsuite/btcd/btcec/v2/ecdsa"
+	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/rpcclient"
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
-	"github.com/btcsuite/btcutil"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/pkg/errors"
 
@@ -153,7 +154,10 @@ func (signer *Signer) AddWithdrawTxOutputs(
 	}
 
 	// 1st output: the nonce-mark btc to TSS self
-	tssAddrP2WPKH := signer.TSS().BTCAddressWitnessPubkeyHash()
+	tssAddrP2WPKH, err := signer.TSS().BTCAddress(signer.Chain().ChainId)
+	if err != nil {
+		return err
+	}
 	payToSelfScript, err := bitcoin.PayToAddrScript(tssAddrP2WPKH)
 	if err != nil {
 		return err
@@ -269,7 +273,7 @@ func (signer *Signer) SignWithdrawTx(
 	}
 
 	// sign the tx
-	sigHashes := txscript.NewTxSigHashes(tx)
+	sigHashes := txscript.NewTxSigHashes(tx, txscript.NewCannedPrevOutputFetcher([]byte{}, 0))
 	witnessHashes := make([][]byte, len(tx.TxIn))
 	for ix := range tx.TxIn {
 		amt, err := bitcoin.GetSatoshis(prevOuts[ix].Amount)
@@ -293,12 +297,11 @@ func (signer *Signer) SignWithdrawTx(
 
 	for ix := range tx.TxIn {
 		sig65B := sig65Bs[ix]
-		R := big.NewInt(0).SetBytes(sig65B[:32])
-		S := big.NewInt(0).SetBytes(sig65B[32:64])
-		sig := btcec.Signature{
-			R: R,
-			S: S,
-		}
+		R := &btcec.ModNScalar{}
+		R.SetBytes((*[32]byte)(sig65B[:32]))
+		S := &btcec.ModNScalar{}
+		S.SetBytes((*[32]byte)(sig65B[32:64]))
+		sig := btcecdsa.NewSignature(R, S)
 
 		pkCompressed := signer.TSS().PubKeyCompressedBytes()
 		hashType := txscript.SigHashAll